+++ Changelog.md
@@ -1,7 +1,13 @@
 Changelog
 =========
 
-0.1.0 (2016-07-06)
+0.1.1 (2016-07-07)
 --------------------
+* [Improvement] Rename cli command current -> last
+* [Improvement] updated append command, added --no-edit key
+* [Improvement] improve add_message feature, new --split-by key to add multiple entries at once
+
+0.1.0 (2016-07-06)
+------------------
 * [Feature] Implemented basic functionality: init, messages, release, append, current, edit